Extract CohortColorSchema from util to separate file
Caleydo/cohort#693
