+++ src/Taskview/tasks/Details.ts
@@ -6,7 +6,8 @@ import { ICohort } from '../../CohortInterfaces';
 import { colors } from '../../config/colors';
 import { IAttribute } from '../../data/Attribute';
 import { getCohortData } from '../../base/rest';
-import { CohortColorSchema, getAnimatedLoadingText } from '../../util';
+import { getAnimatedLoadingText } from '../../util';
+import { CohortColorSchema } from '../../CohortColorSchema';
 import { getIdTypeFromCohort } from '../../utilIdTypes';
 import { DATA_LABEL } from '../visualizations';
 import { ATask } from './ATask';
