+++ src/Taskview/visualizations/AVegaVisualization.ts
@@ -8,7 +8,8 @@ import { Cohort, getCohortLabel } from '../../Cohort';
 import { ICohort } from '../../CohortInterfaces';
 import { IAttribute, IdValuePair } from '../../data/Attribute';
 import { IEqualsList, INumRange, NumRangeOperators } from '../../base/rest';
-import { CohortColorSchema, IFilterDesc, log } from '../../util';
+import { IFilterDesc, log } from '../../util';
+import { CohortColorSchema } from '../../CohortColorSchema';
 import { FilterEvent, SplitEvent } from '../../utilCustomEvents';
 import { Option, VisConfig } from './config/VisConfig';
 import { DATA_LABEL } from './constants';
